Refactored Auton_TaskRunnerDrive to use RobotBaseOpMode
